requirements.txt
fastapi==0.103.1
sqlmodel==0.0.8
uvicorn==0.23.2
aiosqlite==0.19.0
nest_asyncio==1.5.8
argon2-cffi==23.1.0
python-jose==3.3.0
pytz==2023.3.post1
python-multipart==0.0.6
email-validator==2.0.0.post2